Set two variable files from site.yml
That way we can override all the config from a central location. The
`user.yml` file overrides `defaults.yml`. `user.yml` was added to
`.gitignore` in order to allow a user to change it without having to
worry about merges. It's easy to do a diff to see what changed between
the defaults and your own settings.

Needs documentation, this is just a proposal.

Refs #39