Set permissions and owner on virtual mail domains
Without the owner being set to vmail and dovecot, when a virtual user
tries to access their mail account, dovecot throws an error that it does
not have permission to create the mail folder for the user.

With the owner and permissions being set the user's mail directory is
created and they can successfully sign in.