Set dovecot configuration permissions to readable
In order for roundcube to set create a hashed version of the password it
will invoke `doveadm pw`. The `doveadm` program is invoked as the
www-data user which did not have permission to read the dovecot
configuration files causing an error. This change enables the www-data
user to invoke `doveadm`.