Indicate how to make dovecot and znc password hashes in Python
Solves the chicken-or-the-egg problem of generating a dovecot and znc salted
password hash without already having dovecot and znc installed. That is, people
are installing sovereign typically do not have dovecot and znc installed and
have no way of generating the required salted password hashes.

Python 3 standard library generates SHA512-CRYPT salted password hashes by
default so it is particularly convenient for configuring dovecot:

    python3 -c 'import crypt; print(crypt.crypt('password'))'

For ZNC, the required command is slightly more verbose.