Move to Roundcube 1.1.4
Start addressing #456.  Stop using the roundcube package on Debian.
It is outdated on Debian 7 and not supported at all in Debian 8.

Approach is to install by download from roundcube.net and configure.
Roundcube also has a plugin manager now to help automate download
and setup of plugins.

The user is left to finalize the installation (set up the initial
database) and delete the installer.  I am doing this for 1)
idempotence, and 2) to facilitate upgrades to newer versions of
Roundcube.

I am using MySQL for the database server.  I don't know if we
should be using PostgreSQL instead.