Configure roundcube
Modify the configuration files to finalize the install of Roundcube.

The configuration changes are intended to be identical to how Roundcube
was configured on Wheezy.  However, Jessie ships with PHP 5.6.  This
version of PHP enforces SSL certificate checking by default.  This means
using 127.0.0.1 or localhost in SSL connection strings fails.
`{{ mail_server_hostname }}` is used in these places instead.