Correct configuration of roundcube package
The task list was not written correctly.  Exactly one module runs per
task, so each invocation of the debconf module must be done with a
separate task.

The debconf module reports a change when updating a password in the
debconf database even if there is no change to the password.  Set these
tasks to ignore changes.