this will install roundcube from source. closes #8.
this branch will install roundcube from source. tested with 1.2.1.

configuration of roundcube version and plugins is in default.yml.
it will also install configuration for the password plugin but
will not activate the plugin.

the configuration is also present to install and configure skins
from roundcubeskins.net if the user drops a list into user.yml
as follows:
```
roundcube_skins:
  - alpha
  - icloud
roundcube_default_skin: alpha
```