Ansible 2.6 (#773)
* Bump Ansible to 2.6

Older versions are end of life and not receiving security fixes.

* Replace include with import_tasks

Ansible 2.4 introduces include vs import. Use import_tasks for our former task
includes.

Also, the inline tags doesn't get applied properly, not sure if this is a bug in
Ansible.

* linting updates for ansible-lint 4

* Add ansible-lint configuration

* Replace state=installed with state=present

This fixes the deprecation warnings. state=installed is deprecated and will be
removed in Ansible 2.9.

* Pipe syntax for conditional checks is deprecated

* rfc

- use pkg for consistency
- include trusty in unsupported distributions for OwnCloud